Share post metadata between post detail/card/edit views
- Move logic to shared post-metadata directive
- Use this in post card/detail/editor
