+++ app/post/post-module.js
@@ -2,6 +2,7 @@ angular.module('ushahidi.posts', [])
 
 .directive('postCard', require('./directives/post-card-directive.js'))
 .directive('postActions', require('./directives/post-actions-directive.js'))
+.directive('postMetadata', require('./directives/post-metadata.directive.js'))
 .directive('postMessages', require('./directives/post-messages-directive.js'))
 .directive('postAddForm', require('./directives/post-add-form-directive.js'))
 .directive('postValue', require('./directives/post-value-directive.js'))
