+++ server/www/templates/posts/card.html
@@ -6,35 +6,7 @@
     <div class="postcard-body">
         <h1 class="postcard-title"><a ng-href="/posts/{{post.id}}">{{ post.title || post.content | limitTo:150 }}{{ !post.title && post.content.length > 150 ? ' ...' : ''}}</a></h1>
 
-        <div class="metadata">
-            <!-- todo: cleanup / combine this line into a single translation string -->
-            <span ng-show="post.user" class="metadata-author"><img src="http://www.gravatar.com/avatar/{{ post.user.gravatar || '00000000000000000000000000000000' }}?d=retro&s=40" class="avatar">{{ post.user.realname }}</span>
-
-            <span title="{{ displayTimeFull }}">{{ displayTime }}</span>
-
-            <span translate-values="{ source : post.source }" translate="post.source"></span>
-
-            <span class="metadata-visibility public tooltip" ng-show="visibleTo == 'everyone'">
-                <svg class="iconic">
-                  <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#globe"></use>
-                </svg>
-                <span class="bug" translate="post.visible_to_public">This post is visible to the public</span>
-            </span>
-
-            <span class="metadata-visibility tooltip" ng-show="visibleTo == 'draft'">
-                <svg class="iconic">
-                  <use xlink:href="../../img/iconic-sprite.svg#lock-locked"></use>
-                </svg>
-                <span class="bug"  translate="post.visible_to_you">This post is only visible to you</span>
-            </span>
-
-            <span class="metadata-visibility tooltip" ng-show="visibleTo != 'everyone' && visibleTo != 'draft'">
-                <svg class="iconic">
-                  <use xlink:href="../../img/iconic-sprite.svg#lock-locked"></use>
-                </svg>
-                <span class="bug" translate="post.visible_to_roles" translate-values="{ roles : visibleTo }">This post is only visible to Editors</span>
-            </span>
-        </div>
+        <post-metadata post="post" hide-date-this-week="true"></post-metadata>
 
         <div class="postcard-field">
             <post-preview-media post="post"></post-preview-media>
